requirements.txt
instabot
pillow
wand
newspaper3k
sumy
requests
numpy